ingest/requirements.txt
boto3
py-openaq
requests